backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pymongo
python-dotenv
httpx
